refactor : reorder project directory
