+++ algorithms/sort/insertion_sort/src/lib.rs
@@ -54,4 +54,42 @@ where
 }
 
 #[cfg(test)]
-mod tests {}
+mod tests {
+    use std::cmp::Reverse;
+
+    use rand::distr::StandardUniform;
+    use rand::rngs::StdRng;
+    use rand::{Rng, SeedableRng};
+
+    use crate::*;
+
+    const TEST_SIZE: usize = 10_000;
+
+    #[test]
+    fn test_insertion_sort() {
+        let seed: u64 = 42;
+        let rng = StdRng::seed_from_u64(seed);
+
+        let mut vec: Vec<i32> = rng.sample_iter(StandardUniform).take(TEST_SIZE).collect();
+
+        unsafe {
+            insertion_sort(&mut vec);
+        }
+
+        assert!(vec.is_sorted());
+    }
+
+    #[test]
+    fn test_insertion_sort_by() {
+        let seed: u64 = 42;
+        let rng = StdRng::seed_from_u64(seed);
+
+        let mut vec: Vec<i32> = rng.sample_iter(StandardUniform).take(TEST_SIZE).collect();
+
+        unsafe {
+            insertion_sort_by(&mut vec, |a: &i32, b: &i32| Reverse(a).cmp(&Reverse(b)));
+        }
+
+        assert!(vec.is_sorted_by(|&a, &b| { a > b }));
+    }
+}
