+++ algorithms/sort/merge_sort/Cargo.toml
@@ -1,11 +1,7 @@
 [package]
-name = "sort"
+name = "merge_sort"
 version = "0.1.0"
 edition = "2024"
-
-[lib]
-name = "sort"
-path = "./src/lib.rs"
 
 [dependencies]
