+++ algorithms/sort/merge_sort/src/lib.rs
@@ -163,4 +163,38 @@ where
 }
 
 #[cfg(test)]
-mod tests {}
+mod tests {
+
+    use crate::*;
+    use std::cmp::Reverse;
+
+    use rand::distr::StandardUniform;
+    use rand::rngs::StdRng;
+    use rand::{Rng, SeedableRng};
+
+    const TEST_SIZE: usize = 10_000;
+
+    #[test]
+    fn test_merge_sort() {
+        let seed: u64 = 42;
+        let rng = StdRng::seed_from_u64(seed);
+
+        let mut vec: Vec<i32> = rng.sample_iter(StandardUniform).take(TEST_SIZE).collect();
+
+        merge_sort(&mut vec);
+
+        assert!(vec.is_sorted());
+    }
+
+    #[test]
+    fn test_merge_sort_by() {
+        let seed: u64 = 42;
+        let rng = StdRng::seed_from_u64(seed);
+
+        let mut vec: Vec<i32> = rng.sample_iter(StandardUniform).take(TEST_SIZE).collect();
+
+        merge_sort_by(&mut vec, |a: &i32, b: &i32| Reverse(a).cmp(&Reverse(b)));
+
+        assert!(vec.is_sorted_by(|&a, &b| { a > b }));
+    }
+}
