+++ algorithms/sort/quick_sort/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "quick_sort"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+
+[dev-dependencies]
+rand = "0.9"
