+++ algorithms/sort/quick_sort/src/lib.rs
@@ -1,9 +1,5 @@
 mod binary_quick_sort;
-mod insertion_sort;
-mod merge_sort;
 mod ternary_quick_sort;
 
 pub use crate::binary_quick_sort::*;
-pub use crate::insertion_sort::*;
-pub use crate::merge_sort::*;
 pub use crate::ternary_quick_sort::*;
