+++ algorithms/sort/quick_sort/src/ternary_quick_sort.rs
@@ -120,9 +120,14 @@ where
 #[cfg(test)]
 mod tests {
 
+    use crate::*;
+    use std::cmp::Reverse;
+
     use rand::distr::StandardUniform;
     use rand::rngs::StdRng;
     use rand::{Rng, SeedableRng};
+
+    const TEST_SIZE: usize = 10_000;
 
     use crate::ternary_quick_sort::ternary_partition;
 
@@ -157,5 +162,29 @@ mod tests {
         let pivot_pos = ternary_partition(&mut vec);
 
         assert!(is_partitioned(&vec, pivot_pos));
+    }
+
+    #[test]
+    fn test_ternary_quick_sort() {
+        let seed: u64 = 42;
+        let rng = StdRng::seed_from_u64(seed);
+
+        let mut vec: Vec<i32> = rng.sample_iter(StandardUniform).take(TEST_SIZE).collect();
+
+        ternary_quick_sort(&mut vec);
+
+        assert!(vec.is_sorted());
+    }
+
+    #[test]
+    fn test_ternary_quick_sort_by() {
+        let seed: u64 = 42;
+        let rng = StdRng::seed_from_u64(seed);
+
+        let mut vec: Vec<i32> = rng.sample_iter(StandardUniform).take(TEST_SIZE).collect();
+
+        ternary_quick_sort_by(&mut vec, |a: &i32, b: &i32| Reverse(a).cmp(&Reverse(b)));
+
+        assert!(vec.is_sorted_by(|&a, &b| { a > b }));
     }
 }
