+++ yt42/src/lib.rs
@@ -7,4 +7,9 @@ pub mod algorithms {
             pub use heap_on_slice::*;
         }
     }
+	pub mod sort {
+		pub use insertion_sort::*;
+		pub use merge_sort::*;
+		pub use quick_sort::*;
+	}
 }
